[ELY-208] Move the new callback handlers so they can be removed (feedback from code review)
